Задача №1 ДЗ-10
